+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [#548](http://github.com/Nelson-numerical-software/nelson/issues/548): hypot builtin: Square root of sum of squares.
 
+- [#555](http://github.com/Nelson-numerical-software/nelson/issues/555): rsf2csf function: Convert real Schur form to complex Schur form.
+
 ## Previous changelog
 
 [Changelog v0.4.x](CHANGELOG-0.5.x.md)